Move activities to "activities", fragments to "fragments", adapters to "framents.adapters"
