+++ app/src/main/java/com/example/android/spotifystreamer/activities/ArtistTracksActivity.java
@@ -1,24 +1,25 @@
-package com.example.android.spotifystreamer;
+package com.example.android.spotifystreamer.activities;
 
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import com.example.android.spotifystreamer.R;
 
-public class MainActivity extends ActionBarActivity {
+
+public class ArtistTracksActivity extends ActionBarActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_artist_tracks);
     }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
+        getMenuInflater().inflate(R.menu.menu_artist_tracks, menu);
         return true;
     }
