+++ app/src/main/java/com/example/android/spotifystreamer/activities/MainActivity.java
@@ -1,23 +1,26 @@
-package com.example.android.spotifystreamer;
+package com.example.android.spotifystreamer.activities;
 
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import com.example.android.spotifystreamer.R;
 
-public class ArtistTracksActivity extends ActionBarActivity {
+
+public class MainActivity extends ActionBarActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_artist_tracks);
+        setContentView(R.layout.activity_main);
     }
+
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_artist_tracks, menu);
+        getMenuInflater().inflate(R.menu.menu_main, menu);
         return true;
     }
