+++ app/src/main/res/layout/activity_artist_tracks.xml
@@ -1,5 +1,5 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:id="@+id/fragment"
-    android:name="com.example.android.spotifystreamer.ArtistTracksActivityFragment"
+    android:name="com.example.android.spotifystreamer.fragments.ArtistTracksActivityFragment"
     tools:layout="@layout/fragment_artist_tracks" android:layout_width="match_parent"
     android:layout_height="match_parent" />
