+++ app/src/main/res/layout/fragment_artist_tracks.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.example.android.spotifystreamer.ArtistTracksActivityFragment">
+    tools:context="com.example.android.spotifystreamer.fragments.ArtistTracksActivityFragment">
 
     <ListView
         android:id="@+id/artist_track_list_view"
